fix openapi request body examples
